Move compute process target to its own file
